Complete foundations/block/01 exercise
I read the README.md.
I cahnges the properties marked for change to the correct values stated
in the README. The margin-left alignment set to auto used all possible
space to the left since auto distributes equally, but only left margin
had auto, so it made div 3 go as far  right as possible.